test(omo-senpi): split the memorian and in-process suites by behaviour
Ultraworked with [omo](https://github.com/code-yeongyu/oh-my-openagent)
