+++ packages/senpi-task/src/runners/in-process.test.ts
@@ -1,138 +1,14 @@
 import { afterEach, describe, expect, test } from "bun:test"
-
-import { mkdtempSync, rmSync } from "node:fs"
-import { tmpdir } from "node:os"
-import { join } from "node:path"
-
-import { createReadToolDefinition, type CreateAgentSessionOptions, type ToolDefinition } from "@code-yeongyu/senpi"
-
+import { rmSync } from "node:fs"
 import { InProcessRunner, RunnerError } from "./in-process"
-import type { ChildSession, ChildSessionListener, ChildSpec } from "./in-process"
+import { baseSpec, createFakeSession, makeTool, tmpSessionDirs } from "./in-process-child-spec.test-support"
+import type { ChildSession } from "./in-process"
 
-const sampleParameters = createReadToolDefinition(process.cwd()).parameters
-
-function makeTool(name: string, onExecute?: () => void): ToolDefinition {
-  return {
-    name,
-    label: name,
-    description: `test tool ${name}`,
-    parameters: sampleParameters,
-    execute: async () => {
-      onExecute?.()
-      return { content: [{ type: "text", text: "ok" }], details: undefined }
-    },
-  }
-}
-
-type FakeSessionControls = {
-  session: ChildSession
-  resolvePrompt: () => void
-  rejectPrompt: (error: unknown) => void
-  emit: (event: { readonly type: string }) => void
-  steerCalls: string[]
-  followUpCalls: string[]
-  abortCalls: number
-  disposeCount: number
-  lastText: { value: string | undefined }
-  promptCalls: number
-  promptTexts: string[]
-}
-
-function createFakeSession(sessionId = "child-session-1"): FakeSessionControls {
-  const listeners = new Set<ChildSessionListener>()
-  const steerCalls: string[] = []
-  const followUpCalls: string[] = []
-  const promptTexts: string[] = []
-  const lastText = { value: undefined as string | undefined }
-  const counters = { abortCalls: 0, disposeCount: 0, promptCalls: 0 }
-  let settle: { resolve: () => void; reject: (error: unknown) => void } | undefined
-  const session: ChildSession = {
-    sessionId,
-    prompt(text: string) {
-      counters.promptCalls += 1
-      promptTexts.push(text)
-      return new Promise<void>((resolve, reject) => {
-        settle = { resolve, reject }
-      })
-    },
-    async steer(text: string) {
-      steerCalls.push(text)
-    },
-    async followUp(text: string) {
-      followUpCalls.push(text)
-    },
-    async abort() {
-      counters.abortCalls += 1
-    },
-    subscribe(listener: ChildSessionListener) {
-      listeners.add(listener)
-      return () => listeners.delete(listener)
-    },
-    getLastAssistantText() {
-      return lastText.value
-    },
-    dispose() {
-      counters.disposeCount += 1
-    },
-  }
-  return {
-    session,
-    steerCalls,
-    followUpCalls,
-    lastText,
-    promptTexts,
-    get abortCalls() {
-      return counters.abortCalls
-    },
-    get disposeCount() {
-      return counters.disposeCount
-    },
-    get promptCalls() {
-      return counters.promptCalls
-    },
-    resolvePrompt: () => settle?.resolve(),
-    rejectPrompt: (error: unknown) => settle?.reject(error),
-    emit: (event) => {
-      for (const listener of listeners) listener(event)
-    },
-  }
-}
-
-const tmpSessionDirs: string[] = []
-
-function makeSessionDir(): string {
-  const dir = mkdtempSync(join(tmpdir(), "senpi-task-in-process-"))
-  tmpSessionDirs.push(dir)
-  return dir
-}
-
-function baseSpec(overrides: Partial<ChildSpec> = {}): ChildSpec {
-  return {
-    taskId: "task-1",
-    cwd: process.cwd(),
-    sessionDir: makeSessionDir(),
-    depth: 0,
-    parentSessionId: "parent-1",
-    rootSessionId: "root-1",
-    prompt: "do the work",
-    ...overrides,
-  }
-}
+const unhandled: unknown[] = []
+const onUnhandled = (reason: unknown): void => { unhandled.push(reason) }
+afterEach(() => { unhandled.length = 0; while (tmpSessionDirs.length > 0) rmSync(tmpSessionDirs.pop() ?? "", { recursive: true, force: true }) })
 
 describe("InProcessRunner", () => {
-  const unhandled: unknown[] = []
-  const onUnhandled = (reason: unknown): void => {
-    unhandled.push(reason)
-  }
-
-  afterEach(() => {
-    unhandled.length = 0
-    process.off("unhandledRejection", onUnhandled)
-    while (tmpSessionDirs.length > 0) {
-      rmSync(tmpSessionDirs.pop() ?? "", { recursive: true, force: true })
-    }
-  })
-
   test("#given a running child #when steered while the prompt is in flight #then the fake session receives it", async () => {
     const fake = createFakeSession()
     const runner = new InProcessRunner({ createSession: async () => fake.session })
@@ -345,18 +221,6 @@ describe("InProcessRunner", () => {
     })
   })
 
-  test("#given handle construction touches a broken subscription #when start is called #then the created session is disposed before start rejects", async () => {
-    const fake = createFakeSession()
-    const broken: ChildSession = {
-      ...fake.session,
-      subscribe: () => { throw new Error("subscribe failed") },
-    }
-    const runner = new InProcessRunner({ createSession: async () => broken })
-
-    await expect(runner.start(baseSpec())).rejects.toThrow("subscribe failed")
-    expect(fake.disposeCount).toBe(1)
-  })
-
   test("#given a prompt that throws #when the child runs #then a typed failure is recorded, the child stays resident, and no rejection escapes", async () => {
     process.on("unhandledRejection", onUnhandled)
     const cause = new Error("prompt boom")
@@ -389,140 +253,4 @@ describe("InProcessRunner", () => {
     await handle.waitForIdle()
 
     expect(seen).toEqual(["agent_start", "agent_end"])
-  })
-})
-
-describe("InProcessRunner thinking level", () => {
-  test("#given a spec carrying a thinking level #when the child session is created #then the level reaches the senpi session options", async () => {
-    // given
-    let captured: CreateAgentSessionOptions | undefined
-    const fake = createFakeSession()
-    const runner = new InProcessRunner({
-      createSession: async (options) => {
-        captured = options
-        return fake.session
-      },
-    })
-
-    // when
-    const handle = await runner.start(baseSpec({ thinkingLevel: "xhigh" }))
-    fake.resolvePrompt()
-    await handle.waitForIdle()
-
-    // then
-    expect(captured?.thinkingLevel).toBe("xhigh")
-  })
-
-  test("#given a spec without a thinking level #when the child session is created #then no level is forced so senpi keeps its default", async () => {
-    // given
-    let captured: CreateAgentSessionOptions | undefined
-    const fake = createFakeSession()
-    const runner = new InProcessRunner({
-      createSession: async (options) => {
-        captured = options
-        return fake.session
-      },
-    })
-
-    // when
-    const handle = await runner.start(baseSpec())
-    fake.resolvePrompt()
-    await handle.waitForIdle()
-
-    // then
-    expect(captured?.thinkingLevel).toBeUndefined()
-  })
-})
-
-describe("InProcessRunner child system prompt", () => {
-  test("#given a spec carrying a system prompt #when the child session is created #then the minimal loader returns it in place of the default", async () => {
-    // given
-    let captured: CreateAgentSessionOptions | undefined
-    const fake = createFakeSession()
-    const runner = new InProcessRunner({
-      createSession: async (options) => {
-        captured = options
-        return fake.session
-      },
-    })
-
-    // when
-    const handle = await runner.start(baseSpec({ systemPrompt: "# Memorian\n\nYou judge turns." }))
-    fake.resolvePrompt()
-    await handle.waitForIdle()
-
-    // then
-    expect(captured?.resourceLoader?.getSystemPrompt()).toBe("# Memorian\n\nYou judge turns.")
-  })
-
-  test("#given a spec without a system prompt #when the child session is created #then the minimal loader keeps returning none", async () => {
-    // given
-    let captured: CreateAgentSessionOptions | undefined
-    const fake = createFakeSession()
-    const runner = new InProcessRunner({
-      createSession: async (options) => {
-        captured = options
-        return fake.session
-      },
-    })
-
-    // when
-    const handle = await runner.start(baseSpec())
-    fake.resolvePrompt()
-    await handle.waitForIdle()
-
-    // then
-    expect(captured?.resourceLoader?.getSystemPrompt()).toBeUndefined()
-  })
-})
-
-describe("InProcessRunner prompt envelope", () => {
-  test("#given a bare envelope #when the child starts #then the spec prompt is delivered verbatim with no ancestry wrapper", async () => {
-    // given
-    const fake = createFakeSession()
-    const runner = new InProcessRunner({ createSession: async () => fake.session })
-    const prompt = "<memorian-input>\ncandidates and transcript only\n</memorian-input>"
-
-    // when
-    const handle = await runner.start(baseSpec({ prompt, promptEnvelope: "bare" }))
-    fake.resolvePrompt()
-    await handle.waitForIdle()
-
-    // then
-    expect(fake.promptTexts).toEqual([prompt])
-  })
-
-  test("#given no envelope #when the child starts #then the subagent ancestry lines wrap the prompt", async () => {
-    // given
-    const fake = createFakeSession()
-    const runner = new InProcessRunner({ createSession: async () => fake.session })
-
-    // when
-    const handle = await runner.start(baseSpec({ prompt: "do the work" }))
-    fake.resolvePrompt()
-    await handle.waitForIdle()
-
-    // then
-    expect(fake.promptTexts).toHaveLength(1)
-    const prompt = fake.promptTexts[0] ?? ""
-    expect(prompt).toContain("You are running as an omo senpi-task child")
-    expect(prompt).toContain("Task id: task-1")
-    expect(prompt).toContain("Task:\ndo the work")
-  })
-
-  test("#given an explicit subagent envelope #when the child starts #then the ancestry wrapper applies exactly like the default", async () => {
-    // given
-    const fake = createFakeSession()
-    const runner = new InProcessRunner({ createSession: async () => fake.session })
-
-    // when
-    const handle = await runner.start(baseSpec({ prompt: "do the work", promptEnvelope: "subagent" }))
-    fake.resolvePrompt()
-    await handle.waitForIdle()
-
-    // then
-    const prompt = fake.promptTexts[0] ?? ""
-    expect(prompt).toContain("You are running as an omo senpi-task child")
-    expect(prompt).toContain("Task:\ndo the work")
-  })
-})
+  })})
